新增组织树递归查询,常量统一封装

+++ asset-platform-idm/src/main/java/com/asset/mapper/MenuMapper.java
@@ -21,7 +21,7 @@ public interface MenuMapper {
 
     int updateByPrimaryKey(Menu record);
 
-    int addMenuRole(@Param("menuId") Long menuId, @Param("roleId") Integer roleId);
+    int addMenuRole(@Param("menuId") Long menuId, @Param("roleId") Long roleId);
 
     Menu getByPath(String applicationId);
